+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2026-01-12
+
+### Added
+
+- `diffray extends install <url>` now auto-adds URL to config
+- `--global` flag to add extends to global config (`~/.diffray/config.json`)
+
 ## [0.3.1] - 2026-01-12
 
 ### Added
@@ -96,7 +103,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Major codebase cleanup and modernization
 
-[Unreleased]: https://github.com/diffray/diffray/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/diffray/diffray/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/diffray/diffray/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/diffray/diffray/compare/v0.2.0...v0.3.1
 [0.2.0]: https://github.com/diffray/diffray/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/diffray/diffray/compare/v0.1.2...v0.1.3